feat(mget): replaced mgetDoc with DocumentRef
Renamed DocumentID to a more apt DocumentRef as it references a document in a specific location. Once doing that, it made sense to use it for the MGetRequest instead of the unexperted mgetDoc.

This change also lets users to submit their own implementation of RoutableDoc if they are leveraging it.